+++ sql/planar/makeBiconnectedPlanar.sql
@@ -1,0 +1,57 @@
+/*PGR-GNU*****************************************************************
+File: makeBiconnectedPlanar.sql
+
+Copyright (c) 2007-2026 pgRouting developers
+Mail: project@pgrouting.org
+
+Function's developer:
+Copyright (c) 2026 Mohit Rawat
+Mail: mohit25rawat at gmail.com
+
+------
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation; either version 2 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+ ********************************************************************PGR-GNU*/
+
+------------------
+-- pgr_makeBiconnectedPlanar
+------------------
+
+--v4.1
+CREATE FUNCTION pgr_makeBiconnectedPlanar(
+    TEXT,   -- edges_sql (required)
+    OUT seq BIGINT,
+    OUT start_vid BIGINT,
+    OUT end_vid BIGINT)
+
+RETURNS SETOF RECORD AS
+$BODY$
+    SELECT seq, start_vid, end_vid
+    FROM _pgr_makeBiconnectedPlanar(_pgr_get_statement($1));
+$BODY$
+LANGUAGE SQL VOLATILE STRICT
+COST ${COST_HIGH} ROWS ${ROWS_HIGH};
+
+
+COMMENT ON FUNCTION pgr_makeBiconnectedPlanar(TEXT)
+IS 'pgr_makeBiconnectedPlanar
+- EXPERIMENTAL
+- Undirected graph
+- Parameters:
+  - edges SQL with columns: id, source, target, cost [,reverse_cost]
+- Documentation:
+  - ${PROJECT_DOC_LINK}/pgr_makeBiconnectedPlanar.html
+';